Move initialize to new architecture
