+++ packages/db/src/loaders/commands/index.ts
@@ -3,5 +3,4 @@ const debug = logger("db:loaders:commands");
 
 export { generateCompileLoad } from "./compile";
 export { generateNamesLoad } from "./names";
-export { generateInitializeLoad } from "./initialize";
 export { generateMigrateLoad } from "./migrate";
