--- packages/db/src/loaders/resources/projects/add.graphql.ts
@@ -1,12 +1,0 @@
-import gql from "graphql-tag";
-
-export const AddProjects = gql`
-  mutation AddProjects($projects: [ProjectInput!]!) {
-    projectsAdd(input: { projects: $projects }) {
-      projects {
-        id
-        directory
-      }
-    }
-  }
-`;
