+++ packages/db/src/loaders/resources/projects/index.ts
@@ -4,24 +4,9 @@ const debug = logger("db:loaders:resources:projects");
 import { IdObject } from "@truffle/db/meta";
 import { Process } from "@truffle/db/project/process";
 
-import { AddProjects } from "./add.graphql";
 import { AssignProjectNames } from "./assign.graphql";
 import { ResolveProjectName } from "./resolve.graphql";
-export { AddProjects, AssignProjectNames, ResolveProjectName };
-
-export function* generateProjectLoad(
-  directory: string
-): Process<DataModel.Project> {
-  const result = yield {
-    type: "graphql",
-    request: AddProjects,
-    variables: {
-      projects: [{ directory }]
-    }
-  };
-
-  return result.data.projectsAdd.projects[0];
-}
+export { AssignProjectNames, ResolveProjectName };
 
 export function* generateProjectNameResolve(
   project: IdObject<DataModel.Project>,
